Change release mechanic
I've changed the box release mechanic to use Physics2D.IgnoreCollision between the platform and that box. This is way better than what I was doing before where I was turning off the platform collider for a few frames because if you adjusted the speed down, it could turn back on while the box was still falling and cause all kinds of other problems.